requirements.txt
discord.py
python-dotenv
jishaku
pillow
cbvx
tortoise-orm
matplotlib
numpy